+++ CHANGELOG.md
@@ -7,6 +7,15 @@ If it changes significantly or very much, the first digit will move.
 
 ***
 
+## [Beta 1.1](https://gitlab.com/Avespoir_Project/Avespoir/-/tree/Beta-1.1) - 2020-03-25
+
+### Added
+
+* Repository only build constitution
+* Moderator command
+  * DBUserChangeRole Command
+* Bot status display allow users count
+
 ## [Beta 1.0](https://gitlab.com/Avespoir_Project/Avespoir/-/tree/Beta-1.0) - 2020-03-24
 
 Start Beta version!